facto, removed cloudinaryController and created cloudinaryServiceImpl
