--- src/main/java/adeuxpas/back/dto/mapper/ArticlePictureMapper.java
@@ -1,19 +1,0 @@
-// package adeuxpas.back.dto.mapper;
-
-// import org.mapstruct.CollectionMappingStrategy;
-// import org.mapstruct.Mapper;
-// import org.mapstruct.ReportingPolicy;
-
-// import adeuxpas.back.dto.ArticlePictureDTO;
-// import adeuxpas.back.entity.ArticlePicture;
-
-// @Mapper(componentModel = "spring", unmappedTargetPolicy =
-// ReportingPolicy.IGNORE, collectionMappingStrategy =
-// CollectionMappingStrategy.ADDER_PREFERRED)
-// public interface ArticlePictureMapper {
-
-// ArticlePicture articlePictureDTOToArticlePicture(ArticlePictureDTO
-// articlePictureDTO);
-// }
-
-// TO DO : à virer ?
