--- src/main/java/adeuxpas/back/entity/Image.java
@@ -1,86 +1,0 @@
-package adeuxpas.back.entity;
-
-import jakarta.persistence.*;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-
-@Entity
-public class Image {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-
-    private String imageUrl;
-
-    private String imageId;
-
-    public Image() {
-    }
-
-    public Image(String name, String imageUrl, String imageId) {
-        this.name = name;
-        this.imageUrl = imageUrl;
-        this.imageId = imageId;
-
-    }
-
-    /**
-     * @return Long return the id
-     */
-    public Long getId() {
-        return id;
-    }
-
-    /**
-     * @param id the id to set
-     */
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    /**
-     * @return String return the name
-     */
-    public String getName() {
-        return name;
-    }
-
-    /**
-     * @param name the name to set
-     */
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    /**
-     * @return String return the imageUrl
-     */
-    public String getImageUrl() {
-        return imageUrl;
-    }
-
-    /**
-     * @param imageUrl the imageUrl to set
-     */
-    public void setImageUrl(String imageUrl) {
-        this.imageUrl = imageUrl;
-    }
-
-    /**
-     * @return String return the imageId
-     */
-    public String getImageId() {
-        return imageId;
-    }
-
-    /**
-     * @param imageId the imageId to set
-     */
-    public void setImageId(String imageId) {
-        this.imageId = imageId;
-    }
-
-}
