+++ src/main/java/adeuxpas/back/service/CloudinaryService.java
@@ -1,82 +1,10 @@
 package adeuxpas.back.service;
 
-import com.cloudinary.Cloudinary;
-import com.cloudinary.Transformation;
-import com.cloudinary.utils.ObjectUtils;
+import java.io.IOException;
+import java.util.Map;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-
-@Service
-public class CloudinaryService {
-    private static final Logger logger = LoggerFactory.getLogger(CloudinaryService.class);
-    Cloudinary cloudinary;
-
-    public CloudinaryService(
-            @Value("${cloudinary.cloud_name}") String cloudName,
-            @Value("${cloudinary.api_key}") String apiKey,
-            @Value("${cloudinary.api_secret}") String apiSecret) {
-
-        Map<String, String> valuesMap = new HashMap<>();
-        valuesMap.put("cloud_name", cloudName);
-        valuesMap.put("api_key", apiKey);
-        valuesMap.put("api_secret", apiSecret);
-
-        cloudinary = new Cloudinary(valuesMap);
-    }
-
-    @SuppressWarnings("unchecked")
-    public Map<String, Object> upload(String publicId, MultipartFile multipartFile) throws IOException {
-        File file = convert(multipartFile);
-        try {
-            Transformation<?> transformation;
-            if (publicId.startsWith("profilePicture")) {
-                System.out.println("profilePicturetoUpload");
-                transformation = new Transformation<>()
-                        .gravity("face")
-                        .height(208).width(208).crop("fill").quality("auto")
-                        .radius("max").chain()
-                        .fetchFormat("webp");
-            } else {
-                transformation = new Transformation<>()
-                        .height(400).width(600).crop("fill").quality("auto")
-                        .fetchFormat("webp");
-            }
-            Map<String, Object> result = cloudinary.uploader().upload(file,
-                    ObjectUtils.asMap(
-                            "transformation", transformation,
-                            "public_id", publicId));
-            if (!Files.deleteIfExists(file.toPath())) {
-                throw new IOException("Failed to delete temporary file: " + file.getAbsolutePath());
-            }
-            return result;
-        } catch (Exception e) {
-            logger.error("Error uploading file to Cloudinary", e);
-            throw new IOException("Failed to upload file to Cloudinary: " + e.getMessage(), e);
-        }
-    }
-
-    // public Map delete(String id) throws IOException {
-    // return cloudinary.uploader().destroy(id, ObjectUtils.emptyMap());
-    // }
-
-    private File convert(MultipartFile multipartFile) throws IOException {
-        File file = new File(Objects.requireNonNull(multipartFile.getOriginalFilename()));
-        FileOutputStream fo = new FileOutputStream(file);
-        fo.write(multipartFile.getBytes());
-        fo.close();
-        return file;
-    }
+public interface CloudinaryService {
+    Map<String, Object> upload(String publicId, MultipartFile multipartFile) throws IOException;
 }
